test: move nodes command tests to a separate script, test that default peekingduck cli shows help message
