+++ tests/cli/test_cli.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import io
-import math
 import os
 import subprocess
 from pathlib import Path
@@ -24,38 +22,6 @@ from click.testing import CliRunner
 
 from peekingduck import __version__
 from peekingduck.cli import cli
-from peekingduck.declarative_loader import PEEKINGDUCK_NODE_TYPES
-
-PKD_DIR = Path(__file__).resolve().parents[2] / "peekingduck"
-PKD_CONFIG_DIR = PKD_DIR / "configs"
-
-
-def available_nodes_msg(type_name=None):
-    def len_enumerate(item):
-        return int(math.log10(item[0] + 1)) + len(item[1])
-
-    output = io.StringIO()
-    if type_name is None:
-        node_types = PEEKINGDUCK_NODE_TYPES
-    else:
-        node_types = [type_name]
-
-    url_prefix = "https://peekingduck.readthedocs.io/en/stable/nodes/"
-    url_postfix = ".html#module-"
-    for node_type in node_types:
-        node_names = [path.stem for path in (PKD_CONFIG_DIR / node_type).glob("*.yml")]
-        max_length = len_enumerate(max(enumerate(node_names), key=len_enumerate))
-        print(f"\nPeekingDuck has the following {node_type} nodes:", file=output)
-        for num, node_name in enumerate(node_names):
-            idx = num + 1
-            node_path = f"{node_type}.{node_name}"
-            url = f"{url_prefix}{node_path}{url_postfix}{node_path}"
-            node_width = max_length + 1 - int(math.log10(idx))
-            print(f"{idx}:{node_name: <{node_width}}Info: {url}", file=output)
-    print("\n", file=output)
-    content = output.getvalue()
-    output.close()
-    return content
 
 
 @pytest.fixture
@@ -70,6 +36,19 @@ def parent_dir():
 
 @pytest.mark.usefixtures("tmp_dir", "tmp_project_dir")
 class TestCli:
+    def test_help(self):
+        """Checks that calling peekingduck without options and commands prints
+        help message.
+        """
+        result = CliRunner().invoke(cli)
+
+        assert result.exit_code == 0
+        # not testing full message as .invoke() sets program name to cli
+        # instead of peekingduck
+        assert f"Usage:" in result.output
+        assert f"Options:" in result.output
+        assert f"Commands:" in result.output
+
     def test_version(self):
         result = CliRunner().invoke(cli, ["--version"])
 
@@ -77,19 +56,6 @@ class TestCli:
         # not testing full message as .invoke() sets program name to cli
         # instead of peekingduck
         assert f"version {__version__}" in result.output
-
-    def test_nodes_all(self):
-        result = CliRunner().invoke(cli, ["nodes"])
-        print(result.exception)
-        print(result.output)
-        assert result.exit_code == 0
-        assert result.output == available_nodes_msg()
-
-    def test_nodes_single(self):
-        for node_type in PEEKINGDUCK_NODE_TYPES:
-            result = CliRunner().invoke(cli, ["nodes", node_type])
-            assert result.exit_code == 0
-            assert result.output == available_nodes_msg(node_type)
 
     def test_main_py_log_level_debug(self):
         # setup unit test env
